requirements.txt
# =================================================================
# Requirements for Chatterbox Pro
# To install, use: pip install -r requirements.txt
# =================================================================

# --- Core ML & Audio Libraries ---
# Pinned to specific versions for model stability.
# The full URL ensures the correct CUDA version is downloaded.
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.7.0+cu121
torchaudio==2.7.0+cu121
transformers==4.46.3
diffusers==0.29.0
resemble-perth==1.0.1
openai-whisper

# --- GUI Libraries ---
customtkinter==5.2.2
CTkToolTip==0.8
CTkMessagebox==2.7
pygame==2.6.1

# --- Text & File Processing ---
sentence-splitter==1.4
pydub==0.25.1
# For PDF support
pdftextract==0.0.5
# For ePub support
EbookLib==0.18
beautifulsoup4==4.12.3
# For DOCX/MOBI support (also requires system installation of Pandoc)
pypandoc==1.13

# --- Post-Processing (Optional but Recommended) ---
# For audio normalization
ffmpeg-python==0.2.0

# NOTE: The 'chatterbox' library itself is not listed here
# as it is expected to be present as a source folder in the project directory.
# This requirements file handles all its dependencies.